Develop Backpack System
1. Create Backpack SO
2. Create Player SO
3. Create Backpack SO cutom Editor
4. Add Horizontal Layout Group in UI
5. Fix Null Reference bug
6. Create Builder Base

+++ NewMyProject/Assets/02 Scripts/Game/UI/Interface/IUILayoutStrategy.cs
@@ -4,6 +4,6 @@ namespace MyUI.Interface
 {
     public interface IUILayoutStrategy
     {
-        public Vector2 GetPosition(RectTransform parentRectTrans, int x, int y, int width, int height, float spacing);
+        public Vector2 GetPosition(RectTransform parentRectTrans, int x, int y, float width, float height, float spacing);
     }
 }
